Trait ag_iso_stack::driver::Driver

pub trait Driver {
    // Required methods
    fn is_valid(&self) -> bool;
    fn open(&mut self) -> Result<(), DriverOpenError>;
    fn close(&mut self) -> Result<(), DriverCloseError>;
    fn read_nonblocking(
        &mut self,
        frame: &mut Frame
    ) -> Result<(), DriverReadError>;
    fn write_nonblocking(
        &mut self,
        frame: &Frame
    ) -> Result<(), DriverWriteError>;
}

Generic interface for CAN drivers

This layer is meant to abstract the hardware, and should not do its own queuing/buffering.

This trait does not define how to construct and configure a driver, as the details are likely to differ from driver to driver.

Required Methods

fn is_valid(&self) -> bool

Determine whether the driver is connected and healthy

fn open(&mut self) -> Result<(), DriverOpenError>

Open the driver

It is expected you must open the driver after creating it

fn close(&mut self) -> Result<(), DriverCloseError>

Close the driver

It is not necessary to close the driver before dropping it

fn read_nonblocking(&mut self, frame: &mut Frame) -> Result<(), DriverReadError>

Read a Frame from the driver, if possible

This is a non-blocking read. If there is no frame ready to read, this function will return DriverReadError::NoFrameReady.

An out-parameter is used, so that the user can choose whether to construct a new frame for each call, or to re-use memory.

fn write_nonblocking(&mut self, frame: &Frame) -> Result<(), DriverWriteError>

Write a Frame to the driver, if possible

This is a non-blocking write. If the frame cannot be written because the driver’s queue/buffer is full (for drivers like socketcan that do internal buffering), or if it’s otherwise busy, this function will return DriverWriteError::NotReady.

For drivers that defer to some other implementation (Peak, Socketcan), it’s likely that the given frame is copied before being written.